add pin editing
