+++ src/app/(admin)/admin/projects/_components/project-card.tsx
@@ -29,6 +29,8 @@ import { cn } from "@/lib/utils";
 import {Project} from "@/types/projects";
 import {patchProject, postStatus} from "@/hooks/use-projects";
 import {toast} from "sonner";
+import { InlineEditText } from "@/components/forms/inline-text-edit";
+import { InlineStatusSelect } from "@/components/forms/inline-status-edit";
 
 const projectFormSchema = z.object({
     name: z.string().min(1, "Name is required"),
@@ -40,171 +42,6 @@ interface ProjectCardProps {
     project: Project;
     mutate: Function
 }
-
-// Create a reusable InlineEditText component first
-const InlineEditText = ({
-                            value,
-                            onChange,
-                            placeholder,
-                            className,
-                            variant = "input",
-                            rows = 3,
-                            maxLength
-                        }: {
-    value: string;
-    onChange: (value: string) => void;
-    placeholder?: string;
-    className?: string;
-    variant?: "input" | "textarea";
-    rows?: number;
-    maxLength?: number;
-}) => {
-    const [isEditing, setIsEditing] = useState(false);
-    const [tempValue, setTempValue] = useState(value);
-
-    const save = () => {
-        onChange(tempValue);
-        setIsEditing(false);
-    };
-
-    const cancel = () => {
-        setTempValue(value);
-        setIsEditing(false);
-    };
-
-    const handleKeyDown = (e: React.KeyboardEvent) => {
-        if (e.key === "Escape") {
-            cancel();
-        } else if (e.key === "Enter") {
-            // For input: Enter saves
-            // For textarea: Enter only saves with Ctrl/Cmd
-            if (variant === "input" || (e.ctrlKey || e.metaKey)) {
-                e.preventDefault();
-                save();
-            }
-        }
-    };
-
-    if (isEditing) {
-        return (
-            <div className="flex items-start gap-2">
-                {variant === "textarea" ? (
-                    <Textarea
-                        value={tempValue}
-                        onChange={(e) => setTempValue(e.target.value)}
-                        onKeyDown={handleKeyDown}
-                        placeholder={placeholder}
-                        rows={rows}
-                        maxLength={maxLength}
-                        autoFocus
-                        className={cn("flex-1", className)}
-                    />
-                ) : (
-                    <Input
-                        value={tempValue}
-                        onChange={(e) => setTempValue(e.target.value)}
-                        onKeyDown={handleKeyDown}
-                        placeholder={placeholder}
-                        maxLength={maxLength}
-                        autoFocus
-                        className={cn("flex-1", className)}
-                    />
-                )}
-                <div className={cn("flex gap-1", variant === "textarea" ? 'flex-col': '')}>
-                    <Button type="button" size="icon" variant="ghost" onClick={save}>
-                        <CheckIcon className="w-3 h-3 text-green-600" />
-                    </Button>
-                    <Button type="button" size="icon" variant="ghost" onClick={cancel}>
-                        <XCircleIcon className="w-3 h-3 text-red-400" />
-                    </Button>
-                </div>
-            </div>
-        );
-    }
-
-    return (
-        <div className={cn("group flex items-center gap-2 hover:bg-muted/20 rounded-lg px-2 py-1", className)}>
-            <div className="flex-1 cursor-pointer" onClick={() => setIsEditing(true)}>
-                {variant === "textarea" ? (
-                    <div className="min-h-[60px] whitespace-pre-wrap">
-                        {value || (
-                            <span className="text-muted-foreground italic">
-                                {placeholder || "Click to edit"}
-                            </span>
-                        )}
-                    </div>
-                ) : (
-                    <span>
-                        {value || (
-                            <span className="text-muted-foreground italic">
-                                {placeholder || "Click to edit"}
-                            </span>
-                        )}
-                    </span>
-                )}
-            </div>
-
-            <Button
-                type="button"
-                size="icon"
-                variant="ghost"
-                onClick={() => setIsEditing(true)}
-                className="mt-0.5 shrink-0"
-            >
-                <PencilIcon className="w-3 h-3 text-muted-foreground" />
-            </Button>
-        </div>
-    );
-};
-
-// Inline Status Selector
-const InlineStatusSelect = ({ value, onChange, getStatusConfig }: {
-    value: string;
-    onChange: (value: string) => void;
-    getStatusConfig: Function
-}) => {
-    const [isEditing, setIsEditing] = useState(false);
-    const statusConfig = getStatusConfig(value); // Your existing status config function
-
-    if (isEditing) {
-        return (
-            <Select
-                value={value}
-                onValueChange={(newValue) => {
-                    onChange(newValue);
-                    setIsEditing(false);
-                }}
-                open={true}
-                onOpenChange={setIsEditing}
-            >
-                <SelectTrigger className="w-auto">
-                    <SelectValue />
-                </SelectTrigger>
-                <SelectContent>
-                    <SelectItem value="pending">Pending</SelectItem>
-                    <SelectItem value="ongoing">Ongoing</SelectItem>
-                    <SelectItem value="completed">Completed</SelectItem>
-                    <SelectItem value="abandoned">Abandoned</SelectItem>
-                </SelectContent>
-            </Select>
-        );
-    }
-
-    return (
-        <Badge
-            variant="secondary"
-            className={cn(
-                "cursor-pointer hover:opacity-80 transition-opacity",
-                statusConfig.bgColor,
-                statusConfig.color
-            )}
-            onClick={() => setIsEditing(true)}
-        >
-            {value}
-            <PencilIcon className="w-3 h-3 ml-1 opacity-60" />
-        </Badge>
-    );
-};
 
 export function ProjectCard({ project, mutate }: ProjectCardProps) {
     const [dialogOpen, setDialogOpen] = useState(false);
